refactor: gromosXX-compatible CLI and IO cleanup
- Move coordinate reader from md.rs to gromos-io (read_coordinates + CoordinateData)
- Add GENBOX, POSITIONRED, VELOCITYRED block support in coordinate reader
- Rewrite IMD parser for gromosXX positional format (comment headers + data lines)
- Add NTF force flags, COMTRANSROT, shake_tol to ImdParameters
- Refactor md.rs CLI to match gromosXX @-parameters:
  @topo @conf @input (required), @fin @trc @tre @trf @trv (output),
  @pttopo @posresspec @refpos @distrest @angrest @dihrest @gamd (input),
  @verb @print @version @develop (control)
- All simulation parameters now come from @input .imd/.in file
- Fix pre-existing f64/usize type errors in energy_binary and trajectory_binary tests
- 9 new/updated tests for coordinate and IMD parsing
- pair_lj reference still passes at ~1e-11 precision

+++ crates/gromos-io/src/energy_binary.rs
@@ -351,7 +351,7 @@ mod tests {
 
             for i in 0..100 {
                 let mut frame =
-                    EnergyFrame::new(i * 0.002, 100.0 + i, -200.0 - i, 300.0);
+                    EnergyFrame::new(i as f64 * 0.002, 100.0 + i as f64, -200.0 - i as f64, 300.0);
                 frame.bond = -50.0;
                 frame.lj = -100.0;
                 writer.write_frame(&frame).unwrap();
